+++ releaf-i18n_database/spec/lib/releaf/i18n_database/backend_spec.rb
@@ -23,6 +23,22 @@ describe Releaf::I18nDatabase::Backend do
     end
   end
 
+  describe ".locales_pluralizations" do
+    it "returns array all pluralization forms for releaf locales" do
+      allow(Releaf.application.config).to receive(:all_locales).and_return([:de, :ru, :aasdsd])
+
+      allow(TwitterCldr).to receive(:supported_locale?).with(:de).and_return(true)
+      allow(TwitterCldr::Formatters::Plurals::Rules).to receive(:all_for).with(:de).and_return([:one, :other, :many])
+
+      allow(TwitterCldr).to receive(:supported_locale?).with(:ru).and_return(true)
+      allow(TwitterCldr::Formatters::Plurals::Rules).to receive(:all_for).with(:ru).and_return([:one, :other, :few, :zero])
+
+      allow(TwitterCldr).to receive(:supported_locale?).with(:aasdsd).and_return(false)
+
+      expect(described_class.locales_pluralizations).to eq([:one, :other, :many, :few, :zero])
+    end
+  end
+
   describe "#translations" do
     let(:another_translations_store){ Releaf::I18nDatabase::TranslationsStore.new }